simplifies lab-manager route so remove its sub routes and have one table
