+++ app/dashboard/lab-manager/[labId]/delete/page.tsx
