+++ app/dashboard/lab-manager/[labId]/manage/page.tsx
