+++ app/dashboard/lab-manager/[labId]/page.tsx
